1. Minor revision of UI for IIS Open House 2006 demo. 2. Fix a bug of Hsu's keyboard layout detection in OVIMTobaccoOld. 3. Fix a bug of word boundary marking after backspace; the bug was caused by incorrect comparison between size_type and int.

+++ TobaccoOld/PredictorSingleton.cpp
@@ -46,7 +46,7 @@ void PredictorSingleton::clearAll()
 
 void PredictorSingleton::removeWord(size_t position, bool delFlag)
 {
-	size_t tokenIndex;
+	int tokenIndex;
 	if(delFlag)	// true	: DEL
 		tokenIndex = position;
 	else		// false: BackSpace
@@ -62,7 +62,7 @@ void PredictorSingleton::removeWord(size_t position, bool delFlag)
 
 	if(tokenIndex > 0)
 		PredictorSingleton::tokenVector[tokenIndex - 1].withSuffix = false;
-	if(tokenIndex < PredictorSingleton::tokenVector.size() - 1)
+	if(tokenIndex < (int)(PredictorSingleton::tokenVector.size() - 1))
 		PredictorSingleton::tokenVector[tokenIndex + 1].withPrefix = false;
 
 	PredictorSingleton::setTokenVectorByBigram();
